Module lualite::bytecode

Instruction bytecode

All instructions are 32 bits with the first byte being the opcode and the remaining 3 bytes being used for operands. The opcode byte determines how the operand bytes are interpreted.

For the opcode byte, the first 2 bits are the SuperCode. This bit field separates the opcode into 4 categories: arithmetic, comparison, index, and misc.

Modules

constant_value

disassemble

opcode

Representation of the opcode byte

operand

Representation of the operand bytes

Structs

Instruction

Transparent wrapper around an unsigned 32 bit integer

Procedure

Compiled function

Functions

call

cmp_rw

cmp_wr

index

jmp

jmp_if_false

jmp_if_true

math_rw

math_wr

mov

nop

ret