Module lualite::bytecode::operand

Representation of the operand bytes

Structs

ConstantKey

A key into a function’s constant table

FunctionKey

A key into a function table

Global

TODO

Immediate

Integer literal small enough to fit into a single signed byte

InstructionPointer

An index into the bytecode array

RawRegister

Register with a known register number

Enums

WildDestination

Wildcard destination operand (register or global)

WildSource

Wildcard source operand (register, global, immediate, or constant)

Constants

DESTINATION_OFFSET

Bit offset for the destination byte

FIRST_SOURCE_OFFSET

Bit offset for the first source byte

INSTRUCTION_POINTER_OFFSET

Bit offset for an instruction pointer (2 bytes)

SECOND_SOURCE_OFFSET

Bit offset for the second source byte

Traits

AsDestination

Ability for a type to be the destination byte in an instruction

AsSource

Ability for a type to be either of the source bytes in an instruction

FromDestination

Ability for a type to be created from an instruction destination

FromSource

Ability for a type to be created from an instruction source

Register

Types that can be used as a register