Module lualite::parser::expression

Expression parsers

The primary parser to use in this module is expression, which will parse any expression. It’s used in the examples below. The other parsers in this module will only parse a specific type of expression. For example function_call will only parse an expressions where a function call is the root of the expression tree.

Examples:

let source_code = "x + 1.0 * y";
let (_remaining, parsed_ast) = expression(source_code).expect("parse error");

let expected_ast = Expression::Binary {
  left: Box::new(Expression::Identifier(Identifier("x".to_owned()))),
  op: BinaryOperator::Add,
  right: Box::new(Expression::Binary {
    left: Box::new(Expression::Float(FloatLiteral(1.0_f64))),
    op: BinaryOperator::Mul,
    right: Box::new(Expression::Identifier(Identifier("y".to_owned()))),
  })
};

assert_eq!(parsed_ast, expected_ast);

Order of operations can be changed with parentheses:

let source_code = "(x + 1.0) * y";
let (_remaining, parsed_ast) = expression(source_code).expect("parse error");

let expected_ast = Expression::Binary {
  left: Box::new(Expression::Binary {
    left: Box::new(Expression::Identifier(Identifier("x".to_owned()))),
    op: BinaryOperator::Add,
    right: Box::new(Expression::Float(FloatLiteral(1.0_f64))),
  }),
  op: BinaryOperator::Mul,
  right: Box::new(Expression::Identifier(Identifier("y".to_owned()))),
};

assert_eq!(parsed_ast, expected_ast);

Functions

additive_expression

Additive binary operator expressions (+, -)

arg_list

Argument list for a function call

comparison_expression

Comparison expressions (==, !=, <, >=, etc.)

expression

Any possible expression with arbitrary nesting

function_call

Function call expression

index_expression

Indexed container as an r-value

leaf_expression

Expressions with the highest precedence

multiplicative_expression

Multiplicative binary operator expressions (*, /, %)

parenthesized

Parenthesized expression to modify operator precedence order

postfix_expression

Post-fix operators (call and index)

power_expression

An expression raised to the power of another expression

unary_expression

Unary prefix operator expressions